requirements.txt
amqp==5.1.1
asgiref==3.2.10
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-environ==0.4.5
django-timezone-field==5.0
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.0.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
librabbitmq==2.0.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.3.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
xmltodict==0.13.0